added support for non artifact related post execution actions
